# go/completed_task.go
/*
 * Simple ToDo API
 *
 * This is a simple API for managing a TODO List
 *
 * API version: 1.0.0
 * Contact: recruiting@dfsco.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type CompletedTask struct {

	Completed bool `json:"completed"`
}