# go/task.go
/*
 * Simple ToDo API
 *
 * This is a simple API for managing a TODO List
 *
 * API version: 1.0.0
 * Contact: recruiting@dfsco.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Task struct {

	Id string `json:"id"`

	Name string `json:"name"`

	Completed bool `json:"completed,omitempty"`
}