# go/todo_list.go
/*
 * Simple ToDo API
 *
 * This is a simple API for managing a TODO List
 *
 * API version: 1.0.0
 * Contact: recruiting@dfsco.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type TodoList struct {

	Id string `json:"id"`

	Name string `json:"name"`

	Description string `json:"description,omitempty"`

	Tasks []Task `json:"tasks,omitempty"`
}